Initial check-in of NHC, the Node Health Check system.
The Warewulf NHC project provides a script and a framework for
creating and running various system checks on compute nodes to ensure
that they are working correctly.  Though runnable independently or via
cron, the framework is specifically designed for use with the node
health check facility of TORQUE and other resource managers.

Because the execution of the health check may not be threaded, the
Warewulf NHC system takes several steps to minimize the possibility of
hangs or excessive execution time:

 - As much as possible is implemented in native BASH.  The spawning of
   subprocesses (pipes, subshells, etc.) is kept to an absolute
   minimum.
 - Files are read in-place, and filesystem interfaces are used in
   preference to reading command output.
 - A built-in watchdog timer executes in the background to terminate
   the script if it hangs.

The Warewulf NHC system comes with several pre-written checks, and
writing additional checks is easy.  All files in the scripts directory
are automatically included at runtime.

The configuration file syntax allows for string, wildcard/glob-based,
or regular expression-based matching of hostnames; this facilitates
the use of a single configuration file for entire clusters (or even
sites).  A sample configuration is included.

To build RPMs of NHC, simply run "mzbuild" in the nhc directory.  You
will (obviously) need Mezzanine installed.

Feedback to the Warewulf mailing lists is welcome and encouraged!