Add online/offline support for IBM Platform LSF.
NOTE:  This is based entirely on documentation and has not been tested
at all.  If you are an LSF user and are willing to help test, please
contact me!  I haven't found a way to have LSF run NHC on the nodes
yet, so for now it would need to run out of cron or similar.