This should fix the issue spotted by Anthony DelSorbo <adelsorb@csc.com> and Ken Nielson <knielson@adaptivecomputing.com>.
When NHC exits, it tries to terminate the watchdog timer process (a
bash process with a sleep process as a child).  Because of a "bug"
(lack of feature, but that feature is specified by POSIX!) in bash,
its internal implementation of the "kill" builtin is incapable of
sending a signal to an entire process group (see also:  NHC's SVN
r1201 commit).  So when we sent the signal to the watchdog (i.e.,
bash) process, it died, but its child (the sleep process) didn't!

This commit rewrites the watchdog timer (again) to try and make sure
that the sleep goes away when the bash goes away.

NOTE:  This was only an issue if the output of NHC was being piped
(i.e., read()) somewhere.  Unfortunately, that includes pbs_mom....
;-)  You can compare/verify this by running "nhc" by itself on the
command line vs. running "nhc 2>&1 | less"