Add TORQUE/Moab-specific checks.  These are still preliminary, and unit tests for them are still pending, but they're already at least somewhat useful.
 check_moab_sched -t <timeout> -a <alert> -v <version> -m <match>

Checks the output of "mdiag -S -v" against the specified version,
alert, and/or arbitrary match expression(s).  If a matching alert is
found, if the versions don't match, or if any of the match expressions
(possibly negated) trigger, the check fails.  All parameters are
optional.  Multiple occurrences of -m are supported.

 check_moab_rm -t <timeout> -m <match>

Checks the output of "mdiag -R -v" against any specified match
expression(s).  It also looks for any RMs that are not in the "Active"
state.  If any RM is inactive, or if any of the match expressions
(possibly negated) trigger, the check fails.  All parameters are
optional.  Multiple occurrences of -m are supported.

 check_moab_torque -t <timeout> -m <match>

Checks the output of "mdiag -R -v" against any specified match
expression(s).  It also looks for the "scheduling" parameter to be
turned on.  If "scheduling" is false, or if any of the match
expressions (possibly negated) trigger, the check fails.  All
parameters are optional.  Multiple occurrences of -m are supported.