Added new wrapper script:  nhc-wrapper
This wrapper script will execute "nhc"[1] and record the results.  It
then compares the results to the output of the previous run, if
present, and will ignore results that are identical to those
previously obtained.  Old results can be set to expire after a given
length of time (and thus re-reported).  Results may be echoed to
stdout or sent via e-mail.  Once an unrecognized command line option or
non-option argument is encountered, it and the rest of the command line
arguments are passed to the wrapped program intact.

This tool will typically be run via cron(8) and is intended as the
previously promised replacement/enhancement for nhc.cron.  It can be
used to wrap distinct contexts of NHC in a manner identical to NHC
itself (i.e., specified via executable name or command line arg);
also, unlike the existing script, this one does a comparison of the
results rather than only distinguishing between the presence/absence
of output, and results can have a finite lifespan.

Some example invocations:
  To run "nhc" and notify "root" when errors appear, are cleared, or
  every 12 hours while they persist:
    # /usr/sbin/nhc-wrapper -M root -X 12h

  Same as above, but run "nhc -n nhc-cron" instead:
    # /usr/sbin/nhc-wrapper -M root -X 12h -n nhc-cron
      OR
    # /usr/sbin/nhc-wrapper -M root -X 12h -A '-n nhc-cron'

  Same as above, but run "nhc-cron" (symlink to nhc) instead:
    # /usr/sbin/nhc-wrapper -M root -X 12h -P nhc-cron
      OR
    # ln -s nhc-wrapper /usr/sbin/nhc-cron-wrapper
    # /usr/sbin/nhc-cron-wrapper -M root -X 12h

  Expire results after 1 week, 1 day, 1 hour, 1 minute, and 1 second:
    # /usr/sbin/nhc-wrapper -M root -X 1w1d1h1m1s

Some additional enhancements are planned, to potentially extend its
usefulness beyond cron, but this is a good place to start!

Enjoy!

[1] Technically, nhc-wrapper will strip "-wrapper" off the end of its
name and execute whatever remains, or you can specify a subprogram
directly using the -P option on the nhc-wrapper command line.  It was
intentionally written to be somewhat generic in its operation so as to
be potentially useful in wrapping other similar utilities. :-)