nhc:  Fix erroneous reporting of timeouts during final checks.
David Whiteside (GitHub user starboarder2001) noticed (see his PR #9
for details) that NHC would sometimes report that it had timed out
while executing a particular check when, in fact, it had already
completed that check and was trying to finish its work and/or clean up
(e.g., marking a node online).

This tweaks his fix to be even more specific about what exactly NHC is
doing as it goes through the final stages of execution.