test/test_lbnl_file.nhc:  Put all process substitutions on the same lines where they're used.
Some (newer?) versions of BASH don't like creating a process
substitution file on one line and using it on the next while other
(older?) versions seem to handle it just fine.  Keeping everything all
on the same line should work everywhere, so let's do that.  This
should (hopefully?) fix #31 which appears on both Fedora and Cygwin64
but which I had previously misunderstood.