check_hw_mem():  Fix handling of fudge factor
While looking at #60, I noticed that the `check_hw_mem()` function
(unlike its peer, `check_hw_physmem()`), was mishandling fudge factor
values that were passed as absolute size values rather than
percentages.  Specifically, if it saw a percent sign, it would remove
the sign, but it wouldn't multiply that percentage by the amount of
memory in the system until later on...and in a location that was
*outside* the `if` statement that detected the `%` sign.

Fixes #69.