check_dmi_data_match():  Use -! for consistency
As @basvandervlies pointed out in #68, `check_dmi_data_match()` used
`-n` or a bare `!` for negation of the overall check.  The newer
convention is to use `-!` for this, as I mentioned in my prior comment
to that Issue, so this commit updates `check_dmi_data_match()` to
accept that flag.  The prior syntax will be deprecated in a future
version and eventually removed.