test:  Add BASH variable tests for clobbering
Add a new series of sanity checks to make sure, after we've run all
our other tests, that we've not clobbered certain key BASH variables
we rely on for correct operation, debugging, etc.  This is still a
work in progress, so there are more tests to come on this.  But it's a
good start!