scripts/common.nhc:  Fix MAX_SYS_UID auto-detection
While taking another look at @SMark-Black's #50 and #53, I realized
that the code in question regarding `$MAX_SYS_UID` is doing exactly
what it is supposed to do, given the intended meaning of the variable
based on its name.  What was *actually* wrong was that the
`nhc_common_get_max_sys_uid()` function has been reading the wrong
variable!

So `nhc_common_get_max_sys_uid()` will now look for `$SYS_UID_MAX` in
`/etc/login.defs` like it should have been doing all along, and using
the value of `UID_MIN - 1` as a fallback if necessary.

NOTE:  This means that the default auto-detected value of
`$MAX_SYS_UID` will likely be something ending in `99` (like `499` or
`999`) rather than `00` because it was always intended to be (and the
code has always treated it as) the *top* of the exempt UID range, NOT
the bottom of the non-exempt UID range!  If you have any configs or
scripts that rely on different assumptions, please make sure to make
any necessary updates.

Closes #50.