scripts/lbnl_ps.nhc:  Improve LSF support
Based on a couple changes suggested by @SMark-Black in his PR #53, add
another command to look for to auto-detect LSF, and add support for
the LSF `res` daemon to the `check_ps_userproc_lineage()` check.  Also
moved the setting of `$RM_DAEMON_MATCH` to inside the check -- that's
the only thing in that whole entire file that actually requires a
resource manager!