README.md:  Fix Table of Contents
Made a couple tweaks to the `gh-md-toc` script (NHC-specific) to put
the Table of Contents more in line with what the intent of the
formatting was, not necessarily the exact indentation level.  I
intentionally skip heading levels to achieve the correct style, but
that confuses the ToC generator.  Some special casing in the `awk`
script has remedied that.

I haven't taken the time to make the changes in a way that would be
generic enough to consider upstreaming.  Maybe in the future!