nhc:  Reorder RM detection tests so Slurm is first
Since Slurm offers a compatibility package which provides a PBS-style
command-line interface, we need to detect Slurm first to avoid, for
example, mistaking commands from `slurm-torque` for an actual
TORQUE-based cluster.

Granted, this could cause the opposite problem:  Assuming a
Slurm-based cluster if both Slurm and TORQUE/PBS commands are present
when in reality the primary scheduler/RM is TORQUE/Moab or PBS Pro.  I
believe those situations to be, globally speaking, far less
common/frequent; moreover, relying on automatic detection is *always*
prone to erroneous assumptions, the fix for which continues to be
taking the guesswork out of the equation altogether!

But the automatic detection needs to work more often than not, and
this change should help that.  Fixes #20 from @OleHolmNielsen.