nhc:  Don't use "/usr/libexec" et al. if missing
Not all Linux distributions use "/usr/libexec" as it is marked
optional by the FHS.  This adds checks to the `nhcmain_init_env()`
function to make sure that NHC only uses directories that exist.

So now `$SYSCONFIGDIR` defaults to "/etc/sysconfig" ("/etc/default" on
Debian), but if that directory doesn't exist, NHC will fall back on
"/etc/default" if it exists or "/etc/nhc/sysconfig" if not.

Similarly, `$LIBEXECDIR` will fall back to "/usr/lib" if
"/usr/libexec" is not present.

This fixes #75, and it's highly likely to interfere with the ability
to merge #63 at a later date.