nhc:  Activate lazy creation of /etc/nhc/sysconfig
The new fallback for "/etc/sysconfig" is a directly which will rarely
be needed, but if/when it *is* necessary, this change allows NHC to
create the directory lazily on the fly.  And by adding that directory
as a `%ghost` to the spec file, we ensure that RPM will handle it
correctly if and when necessary.