lbnl_ps:  Exclude system UIDs from resource checks
The `nhc_ps_check_res()` function that back-ends the 4 resource
consumption checks previously failed to account for system users
vs. normal users.  It will now default to excluding low-UID system
users (below `$MAX_SYS_UID`) but will also accept a `-A` option to
revert to the old behavior.

This commit also fixes some line/variable sorting and renames the
`$ALL` variable to `$ALLPROC` to make room for the new `$ALLUSERS`
variable (tied to the `-A` option).