lbnl-nhc.spec.in:  Special-case tagged release
If the current version being built (as specified in `configure.ac` and
assigned to `%{version}`) exactly matches the tag for that same
version in the Git repository, swap out the normal Git-based `%{rel}`
string for simply "1" if the working directory is clean or "1.1" if
it's dirty.  (The disttag is, of course, appended in all cases.)

Additionally, since the Git-fu that generates all this stuff isn't
exactly obvious, I added several detailed comments to describe what
each macro is doing and why.