helpers:  Recognize & ignore new "planned" state
Slurm 21.08 has added a *very* welcome new feature called the
"planned" state; this state is used to indicate nodes which would
otherwise appear "idle" but are being held for a large/high-priority
job.  This is a frequent source of user confusion and angst (because
they would look at the node counts and wonder why their job wasn't
running even though enough nodes appeared to be empty), so having this
is a big win for user services/consultant staff and scheduler admins
everywhere!

As with the handling of the `reboot` states, full recognition of this
state will be available in 1.4.4 and 1.5, but for now, the helpers
will just ignore this state completely.