nhc-wrapper:  Allow wrapping of non-system items
The intent was always to allow `nhc-wrapper` to be used for any
arbitrary command, either a system command the user has permission to
run or a custom utility in the same location as the `*-wrapper` link.
This can be seen in the old `nhcwrap_finalize_env()` attempting to
extract the `dirname` and `basename` of `$SUBPROG`.

However, that never worked; the value of `$0` was already split into
`dirname`/`basename` in `nhcwrap_init_env()`...and the former was
discarded completely.  Only the `basename` portion was saved to
`$SUBPROG`.

This fixes the problem by saving both halves of the split value in
`nhcwrap_init_env()` and prepending the `dirname` part to `$PATH`.
After parsing the command line, in case the (undocumented) `-P` option
was provided, `nhcwrap_finalize_env()` checks `$SUBPROG` against the
original `$0`, and if `$SUBPROG` has been altered, it repeats the
above actions with the new value.