nhc:  Tweak the implementation per MR discussion
Based on the discussion in !54, these changes adjust the `syslog()`
message leader to be more customizable and have the general format of
`IDENT[PID]{TAG}`, where
- `IDENT` is either `nhc` or `nhc(CONTEXT)`
- `{TAG}` is omitted if `$SYSLOG_TAG` is unset/empty (the default)
- Facility, level, priority, size limit, identity, and the overall
  command are all independently customizable

Also, the `LOG_SUCCESS` setting, rather than being altered/renamed,
has been removed altogether.  After giving it a lot of thought, the
default behavior has been changed to always `syslog()` on exit, be it
successfully or otherwise.  The log message -- which goes to both the
system log (always) and the NHC log (verbose mode only) -- was also
made more informative.