test/:  Correct/work around new -x weirdness
I'm not sure what changed here, or if I'm just remembering
incorrectly, but it seems that the current BASH behavior on RHEL7 when
in `-x`/`xtrace` mode is to error out completely -- claiming a `Syntax
Error`, even! -- when encountering the `<(...)` / `>(...)` constructs
used in process substitution.  I'm not sure why/how it got *worse*
than it was before, but it's all kinds of frustrating!

For the time being, I'm just skipping the `lbnl_file.nhc` unit tests
altogether.  Not ideal, to say the least!

I also discovered an off-by-one error in one of my skip counts, so I
fixed that too while I was at it.