nhc:  Add support for version flag without ".in"
This is my proposal for how to implement a "version" option for `nhc`
without having to turn it into `nhc.in`.  I think it's important to
maintain the ability to run `./nhc ...` in a Git clone and have it do
something intelligent.  So this adds a "post-install transform" step
to the normal Automake rules that updates the installed `nhc` in-place
with the full version+release string.

Any thoughts?