nhc:  Keep commit hash out of Git's nhc
Somehow, somewhere along the line, the version of the `nhc` script in
the upstream source repositories (e.g., GitHub) got transformed to
include a Git commit hash and RPM release string in it.  By design,
that change is an in-place alteration that should only ever impact the
`nhc` file that's being/been installed on a host.

Rather than fixing it manually, it occurred to me that automating and
enforcing that state would also help me to not forget to change the
`NHC_VERSION` line in `nhc` when I update the canonical version string
(as supplied to `AC_INIT()` in `configure.ac`).

So I added a quick stanza to `configure.ac` -- shockingly similar to
the one in `Makefile.am`, as it happens -- that will detect and
correct the aforementioned unfortunate situation if ever it should
recur!