nhc:  Print check numbers in verbose mode (#124)
When the `VERBOSE` setting is activated, NHC displays each check line
just before running it.  This adds a counter showing the current check
number as well as the total number of checks to be run.  Leading
zeroes (should these be spaces instead??) are also added so that all
lines "line up" properly from top to bottom.