lbnl_file.nhc:  Rethink `stat`/`test` checks
This commit does a number of things:

- Support was added in `check_file_stat()` for handling (almost) every
  portable field that can be selected from `stat -c`; specifically:

   - There's now a separate pair of options for validating just the
     *permissions* bits of the file mode (i.e., `mode | 00007777`) or
     the entire `mode` value (including the higher-order "file type
     bits," `mode | 00170000`).

   - Another new pair of options, `-b`/`-B`, was added to test
     whether or not the "birth timestamp" is older or newer,
     respectively, than the given number of seconds (only on
     filesystems which support it, of course).

   - On Linux hosts, SELinux file contexts are also supported.

- The `check_file_test()` shell function's support for command-line
  arguments was rewritten to use `getopts` like many other checks
  already do.  Among other things, this allows options to be bundled
  together.  This might seem trivial, but being able to turn
  `check_file_test -f -G -O -e -r -w -s ~/.*profile ~/.*rc` into
  `check_file_test -GOefrsw ~/.*profile ~/.*rc` is definitely not
  nothing.

- Variables have replaced a hard-coded format string when executing
  the `stat` command.

- The quality of the unit tests for both of these checks has been
  substantially improved.

And of course there were other tweaks here and there as well.

This is not to say they're complete, though!  As stated in this
commit, the `stat` stuff needs to be refactored to use hashmaps
instead (or "associative arrays" or "dictionaries," if you prefer) now
that the older versions of Bash are no longer supported.  (In fact,
there are several items that need to be "modernized" to use newer Bash
features...!)  But I'd like to get 1.5 released sooner rather than
later, so some features and improvements will be postponed until the
1.5.1 development cycle.