nhc:  Make `flush` an option instead of a function
This has been on my "TODO" list for ages now, so since I'm already
making I/O and logging changes, I decided to include this too.

Prior to this change, `syslog()` would only append message lines to a
string variable; nothing would actually be written to the `syslog`
socket (using the `logger` command by default) until the
`syslog_flush()` function was invoked.

As with many things in NHC, this was done to minimize context switches
and subcommand execution as much as possible, especially during the
"running checks" phase.  There are, however, numerous instances in
which it's prudent to write information to the `syslog` right away
rather than saving it for later; this is made quite evident by the
number of times a call to `syslog_flush()` immediately follows
`syslog()` invocation in NHC's code.  The repetition makes that code
jump out at you; in my humble opinion, it's just too distracting.

This commit adds support to `syslog()` for requesting immediate
flushing of the `syslog` data by specifying the `-f` option.  All
references to `syslog_flush()` have been replaced with `syslog -f`
and, if they followed a call to `syslog()`, merged the two calls into
one.  The `syslog_flush()` function is now deprecated and may be
removed in a future release; however, for compatibility reasons, the
`syslog_flush()` function does still exist and will still do the same
thing it did prior to this commit.  It is now a simple wrapper around
a call to `syslog -f`.

(Another possible eventual change:  If `syslog()` ever acquires any
additional options apart from just `-f`, I will convert it to use
`getopts` like other functions.  For now, though, since it's the only
available flag and takes no argument(s), it's faster and simpler to
just look at the first argument...either it'll be `-f`, or it won't.)