lbnl_file.nhc:  Handle wonky SELinux supportitude
On Linux hosts with fully functional SELinux support in the `disabled`
state, the `stat` command returns some error messages complaining
about a lack of context data:

```
/usr/bin/stat: failed to get security context of '/my/file': No data available
```

On the other hand, a system with SELinux enabled and active will
return the correct value (e.g., `system_u:object_r:null_device_t:s0`
for `/dev/null`).  To address this problem, at least for the short
term, the `%C` for the SELinux Context field will only be included if
SELinux appears to be supported *and* active.  On RHEL-compatible
systems (and possibly others), this means `SELINUX` being set to
something neither empty nor `"disabled"`.  The return value of
`selinuxenabled` is also checked.  For now, that's it, but please do
suggest other methods you know of!

Interestingly, `stat(1)` seems to output each of the values requested
one-by-one in order -- meaning that, even though the error is written
to `stderr`, it's written at the correct point to correlate when that
field would have been output in a non-failure scenario.  In other
words, if I use `'%n %m %C %u %g'` as my `stat -c` format string
(that's "name mountpoint context uid gid") and redirect `stderr` to
`stdout`, the output will contain the file's name in column 1, its
mountpoint in column 2, and column 3 will contain the error message
shown above followed by a newline.  After the newline will come a
space, the numeric UID, another space, and finally the numeric GID.

The point being, all the other information `stat(1)` provides, even if
it gets no SELinux data back, is still valid and still shows up in the
correct columns/fields per the requested format.  So the possibility
exists to better handle this case in the future -- more of a "best
effort" approach rather than "non-zero exit means failure."  Another
approach would be to make it an admin-controlled setting.  I can't
really take the time to do that just now, though.