nhc:  Add support for flag to show version/info
A previous commit (63c5678d) added the `NHC_VERSION` variable,
transformed at install-time, to provide the ability to identify the
version of `nhc` being executed.  This commit adds the `-V` option to
allow that information to be retrieved from the command line or a
shell script.

I'm still working out the specifics of what gets displayed by `-V`, so
the exact output is subject to change, but the basic idea is this:
`-V` displays just the name and version string; for each additional
`-V` added to the `nhc` invocation, additional info/details may be
provided.  For right now, though, I'm keeping it pretty simple.