nhc:  Add new `-VVV` response and tidy up previous
This updates the code for handling `nhcmain_version_info()` invocation
with a count/argument of 0 or 1 (`-V`) to match the style of the
2-count (`-VV`) handler; it also adds support for a value of 3
(`-VVV`) to obtain a reasonably comprehensive view of configuration
variables/settings.

The existing behavior for `-V` and `-VV` was not altered; specifying
`-V` twice will still display the information that a single `-V`
displays followed by some additional details.  Likewise, adding a
third `-V` still displays everything that was included in the output
for the first and second, followed by even more detailed information.
The output is additive, and I feel that the new layout/formatting of
the code makes this more clear and easier to follow.  Conceptually, it
maps well to similar functionality in other software (e.g., the `-v`
flag in OpenSSH).