lbnl_fs.nhc:  Minor adjustment to mount check
A small issue revealed itself during pre-release testing:  When a
mountpoint does not exist, any attempt to mount that mountpoint's
filesystem will, obviously, fail.  For that reason, NHC will
automatically create (or *attempt to create*) any non-existent
mountpoints it encounters for which the admin has specified a check.

(No option/flag was added to control this behavior; logically, if a
directory has a mount configuration for it specified in `fstab` *and*
the admin has configured NHC to verify that a filesystem is mounted
there, it stands to reason that the admin would want a filesystem to
be able to be successfully mounted there...and for that to happen, the
directory would have to exist.)

Previously, NHC would report that the missing directory issue had been
"`auto-fixed`" as part of the result message -- but it wouldn't
actually check to make sure the `mkdir` command had really worked!
And if it *had* failed, the `mkdir` command would usually report the
failure immediately prior to displaying the aforementioned claim of
success, resulting in something like the following:

```
  Running check 06 of 89:  "check_fs_mount -t "nfs" -s "netapp:/moo" -f "/mnt/remote/moo""
  mkdir: cannot create directory '/mnt/remote/moo': No such file or directory
  Health check failed:  check_fs_mount:  /mnt/remote/moo not mounted; directory /mnt/remote/moo missing (auto-fixed)
```

So the admin would see the `mkdir` failure and then NHC claiming
success...every time it ran.  Not exactly confidence-inspiring, right?

This simply tests for the existence of the mountpoint directory, and
if it doesn't exist, reports the auto-fix failure rather than assuming
success.  It's a little thing, perhaps, but the potential for user
confusion was, nonetheless, unquestionably present.