nhc:  Remove unnecessary braces for readability.
Lately I have tried to get myself into the habit of *always* typing
shell/env variables with the often-optional braces; I find this helps
with readability -- especially with color syntax highlighting -- and
is less error-prone.  However (and I'd welcome opinions on this!), I
have not been doing that with the NHC codebase; I saw no need to
inflict my personal neuroses on others, and to be fair, it's not
always a win....

This is one of those times.  Some of the braces in that statement are
necessary -- specifically, the ones doing the conditional use of the
`SYSLOG_TAG` setting, iff[1] it's set -- but fully half of them
weren't.  And since some of the ones that weren't required were mixed
in with the ones that were, it made the logic and syntax more
difficult to follow.

Minor, I know, but maintainability is important!

[1]:  For many years I've known and used "if__f__" (i.e., `iff`, the
word "if" with another "f" added to the end) as a convenient and
efficient shorthand for "if, and *only* if," a phrase frequently used
in mathematics, logic, critical thinking...and, of course, computer
science/engineering.  But it occurred to me just now that it's
possible some devs or end-users might not recognize it as an
abbreviation, or might even mistake it for a typo.  I'm particularly
concerned about folks in other countries.  Is it a universal thing? or
am I just confusing people?

As always, if you have any thoughts or feedback, I welcome all!