nhc:  Sync with latest from messaging subproject
Separately from NHC, I've been doing ongoing work to try to
streamline, simplify (for the consumer, anyway...), and most
importantly to unify as much of the code as possible for the various
styles of messaging we do -- both in NHC and in shell scripts in
general.

This isn't ready for actual use in production, and I may take it out
completely before release; this is just me trying to keep things as
sync'd up as I can so I don't drive myself completely bonkers.