nhc:  Add "pickdir()" function and megacomment
For quite awhile now, I've noticed a frequent need for the same bit of
"useful functionality," a utility routine in the grand tradition of
`execvp()`, but more generalized, and with greater flexibility.
Specifically, `pickdir()` -- as the name implies -- picks a directory
out of a list of candidate directories using a "first-fit" methodology
(as opposed to a "best-fit" algorithm...though maybe I'll add that
later?) by applying a caller-defined conditional to each and returning
the first candidate for which the conditional evaluates to `true`.

The summary comment for `pickdir()` contained in this changeset is
quite thorough, so I'll try to just summarize.  (For all the details
and specifics, peruse the comment right above the definition for
`pickdir()` in the updated `nhc` script.)

`pickdir()` can be invoked in several ways; the only strict
requirement is that it needs at least 1 candidate directory,
preferably a list of them, from which to choose.  The caller can
additionally supply any/all of:  a variable name in which to store the
result (rather than just `echo`ing it to `stdout`, which is the
default behavior); a delimiter character to be used for splitting a
single commandline argument into a list of candidates (e.g., `:` for
`PATH` directories, or `,` for the list of users per group in
colon-column 4 of `/etc/group`); and/or a custom conditional against
which to test each candidate in turn, in search of the first one that
makes the conditional evaluate to a "true" value (specifically, in
this case, `0`).

The use cases for it will become more apparent in upcoming commits for
code that actually calls it! :-)