nhc:  Handle Bash-produced exit codes differently
I recently noticed an issue one of our team ran into when deploying a
new check, and the corresponding config file updates to use it, to a
large swath of nodes, whereupon every single one of the aforementioned
nodes was promptly drained by `nhc` immediately upon its next run.

Unfortunately, I wasn't able to help; I was reading through the
Slack/Mattermost discussion after the problem had already been
resolved.  But I noticed that a few people had some difficulty
interpreting what NHC was trying to tell them.  I saw it right away,
but only because I eat, sleep, and breathe this stuff!  The answer
simply wasn't going to jump out at anyone without a lot of
shell-scripting experience.

Long story short(er), the problem turned out to be that the new check
(call it `check_moo_volume()`) was in a new script file (call it
`lanl_moo.nhc`), and `lanl_moo.nhc` was not pushed out to these nodes
at the same time the config updates were, so `nhc` was receiving exit
code `127` back when it tried to run `check_moo_volume ouch 120dB` on
the node.

The reason it stood out to me was simply that I instinctively know a
`127` return code means, essentially, "command not found."  And what's
the "command" that's "not found?"  For `nhc` running checks, it's the
check itself!  And if the check's not found, that means the check
function isn't defined, thus the place where the function definition
should be is missing or unusable.

Once you recognize that `127` means "missing check/script," it doesn't
take long to piece together what happened.  This commit is intended to
help reduce the time from incident to recognition to understanding to
resolution by providing special error messages ("Reasons" in
Slurmspeak) that make it clear(er) what happened and nudge the
admin/user toward the most likely causes.

If you have any suggestions for other exit codes or failure scenarios
that NHC could help with recognizing, please file an Issue on
GitHub/GitLab (as appropriate) or reach out via e-mail!