nhc:  Promote "none" resource manager to 1st class
This change does a couple things.  For one, and probably of most
interest to people, is that `nhc` now supports the `-r` option; you
can use it to specify your desired RM without having to modify configs
or manipulate the environment.  Second, use of `none` as the value for
`NHC_RM` wasn't fully baked, but it is now fully recognized and
handled just like any other supported resource manager.

One really nice addition that has come out of this:  specifying "none"
as your RM will now automatically disable `MARK_OFFLINE` as well as
the online/offline helpers, so end-users no longer have to do it
manually (and *remember* to do it each and every time)!