nhc:  Clean up and enhance hostname value handling
One of NHC's key features is its ability to prevent future jobs from
starting on nodes that fail their health check(s); to do this, it must
contact the appropriate scheduler or resource manager (RM) responsible
for scheduling work on that node and direct it to pull the node out of
service.  This, in turn, requires that NHC reference the node the same
way that the RM/scheduler does.

This is usually not an issue because, in most cases, the 2 values are
identical.  Some platforms -- like Crays, for example -- have multiple
simultaneous names for each node, and the "best" value to use for each
case may or may not match.

I've often been asked the best way to handle such a situation, and
until now, I haven't really had a good, clean, robust answer to
provide.  Now I do.  The config variable `NHC_HOST` and the `-m
<name>` argument are now available to clearly specify how NHC should
refer to the current node/host/system.

NOTE:  This same value is also used for the match string at the start
of each config file line, so changing this setting may also impact
which checks are run.  Be sure to make necessary changes to the config
file, if any, when changing this value!

Fixes #129, albeit with some significant changes to the original plan!