nhc-test:  Tests must be deterministic...
I had forgotten the test I wrote to make sure `nhc`'s help output
stays within the standard default 80-column terminal width.  When I
added the new `-m` option and included `$HOSTNAME` in the help output,
I failed to realize that I had broken the determinism of that
particular test by allowing it to use the `HOSTNAME` value of whatever
host was running the tests.

So naturally, it wasn't until I ran the tests on a system with an FQDN
as its UTS hostname that I saw the early test failure suddenly
appear....

Simple solution, of course:  hard-code the `HOSTNAME` value to
something short.  Determinism restored.  World saved.  That's lunch.