nhc:  Add/correct new path calculations and vars
Once again, my colleague @papamoose did me a solid and gave my latest code a
run-through...and it's a good thing too.  He caught some things that I'd
completely missed, most notably that I'd __hard-coded__ the initial path of
`/var` for some of the settings, and as a result, I'd never be able to reach
my goal of simplifying the user experience!  The only ways to override those
paths permanently are the exact same ways I was trying to avoid!

To be more specific, before all these changes, the user would have to either
 1.  Pass the specific set of necessary command-line options/arguments to
     "fix" the directory locations, and do so each and every time `nhc` is
     executed.  For example, perhaps something like this:
     `nhc -D ~/.nhc/etc LIBEXECDIR=~/.nhc/libexec`
     `INCDIR=~/.nhc/scripts HELPERDIR='' ...`
     and so on.  You get the idea.
 2.  Alter NHC's config settings in `~/.bashrc` and pray for no conflicts
 3.  Edit the `nhc` driver script itself to change the default paths.

None of these options is acceptable to me.  Now, with the changes in the
prior commits (c960f8c1 through 4fcf83e9), the user could set the
`NHC_CFG_GLOBAL` environment variable in their `~/.bashrc`, place all the
other settings changes they want into that file, and be done; *or*, at the
user's option, they could build NHC with `--prefix="~/moo"` on their
`./configure` (or `./autogen.sh`) command line, do `make install`, and add
the new location of `nhc` (`~/moo/sbin`) to their `PATH`.  The rest should be
discernable by `nhc` without requiring any other actions by the user.

Anywhy...  To address the problems @papamoose pointed out, I added 2 new
config variables, `LOGDIR` and `RUNDIR`, that each build off of the
`STATEDIR` variable, also new, which is a direct analogue to the
`--localstatedir` value passed to `./configure` in GNU Autotools and to the
`%{_localstatedir}` standard RPM macro.

The new settings are handled like the prior ones, and I added the
respective `NHC_CFG_whateverDIR` variables as well.  As part of this work, I
also redid some of the code in `nhcmain_finalize_env()` and elsewhere that
handles post-config synchronization of settings and/or now use these new
variables.

I also tweaked `nhcmain_load_sysconfig()` and `nhcmain_parse_cmdline()` so
that the appropriate file(s) in `$SYSCONFIGDIR/` will be reloaded if their
location was changed on the command line.  I say "file(s)" rather than "file"
because I *also* put in support for a per-context global config file,
`$SYSCONFIGDIR/$NAME`, to the loaded immediately after `$SYSCONFIGDIR/nhc`
itself.  And finally, a non-fatal warning was added, and the problem worked
around, if the location for the PID file was inaccessible to `nhc`.

Unit tests were updated in sync with the above changes, but a number of
additional unit tests still need to be written to fully exercise all the new
directory path calculation combinations, especially with respect to fallback
and last-resort handling.  But the production team's needs matter more than
my wants!